perf: introduce a11y probe to avoid rerendering
